Move dr_flac/dr_flac.h to dr/dr_flac.h
